% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averageCurve.R
\name{averageCurve}
\alias{averageCurve}
\title{Calculate an average curve from a list of fits, using the AIC weights.}
\usage{
averageCurve(fitList, xPredict)
}
\arguments{
\item{fitList}{List of fitted models (from fitModels())}

\item{xPredict}{Vector of values where the curve should be estimated}
}
\value{
A data frame with x and y columns
}
\description{
Calculate an average curve from a list of fits, using the AIC weights.
}

