% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildModel.R
\name{buildModel}
\alias{buildModel}
\title{Build a model object used for model fitting later on}
\usage{
buildModel(mFun, mName, mFormula, mParams, mStarting)
}
\arguments{
\item{mFun}{Function of the form f(x, params), where params is a named
list of parameters and x a vector of input data}

\item{mName}{String, name of the model}

\item{mFormula}{Formula describing the model}

\item{mParams}{Vector of strings containing the parameters names}

\item{mStarting}{Named list with default starting values when fitting
the model to some data}
}
\value{
An object of class "customModel" (basically a list)
}
\description{
Build a model object used for model fitting later on
}
\examples{
# Simple linear model

f = function(x, params) {
      a = params[["a"]]
      b = params[["b"]]
      return(a + b * x)
    }

myModel = buildModel(f, "linear model",
            y ~ a + b * x, c("a", "b"),
            list(a = 1,  b= 1))

x = c(1, 1, 2, 3, 4, 4, 5, 7)
y = c(2, 4, 3, 7, 8, 7, 10, 11)

fit = fitModel(myModel, x, y)

}
