% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateAIC.R
\name{calculateAIC}
\alias{calculateAIC}
\title{Calculate AIC and AIC weights for a list of fits}
\usage{
calculateAIC(fits)
}
\arguments{
\item{fits}{List of fit results, e.g. output from fitModels()}
}
\value{
Data frame with class "fitAIC"
}
\description{
Calculate AIC and AIC weights for a list of fits
}
\examples{
x = c(23, 25, 27, 29, 31, 32, 33)
y = c(0.1, 0.25, 0.3, 0.6, 0.8, 0.6, 0.1)
fits = fitModels(getModelLibrary(), x, y)
calculateAIC(fits)

}
