% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCT.R
\name{findCT}
\alias{findCT}
\title{Find the low and high critical temperatures (CT_low and CT_high) for
which the average performance curve built from fitList crosses the
threshold fractionOptimum * optimumPerformance.}
\usage{
findCT(fitList, interval, fractionOptimum)
}
\arguments{
\item{fitList}{List of fits, output from fitModels()}

\item{interval}{Vector determining the range in which CTs are searched
for (e.g. c(15, 40))}

\item{fractionOptimum}{Numeric between 0 and 1}
}
\value{
Named vector
}
\description{
Note: No check is done to ensure that those CTs are unique. The curve
should be checked visually to ensure that the analysis makes sense.
}

