% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModel.R
\name{fitModel}
\alias{fitModel}
\title{Fit a previously built model on some experimental data}
\usage{
fitModel(customModel, x, y, initParams = NULL)
}
\arguments{
\item{customModel}{Model built with buildModel()}

\item{x}{Numeric vector, x data}

\item{y}{Numeric vector of same length as x, y data}

\item{initParams}{Named list of starting values for model fitting, if
NULL then the default values specified at model building are used}
}
\value{
Output from nlsLM
}
\description{
The fitting itself is performed with the nlsLM() function from the
minpack.lm package.
}

