% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModels.R
\name{fitModels}
\alias{fitModels}
\title{Fit a list of previously built models on some experimental data}
\usage{
fitModels(customModels, x, y, initParams = NULL)
}
\arguments{
\item{customModels}{List of models built with buildModel()}

\item{x}{Numeric vector, x data}

\item{y}{Numeric vector of same length as x, y data}

\item{initParams}{List of named lists of starting values for model fitting, if
NULL then the default values specified at model building are used}
}
\value{
Output from nlsLM
}
\description{
The fitting itself is performed with the nlsLM() function from the
minpack.lm package.
}
