% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolateAtTemp.R
\name{interpolateAtTemp}
\alias{interpolateAtTemp}
\title{Interpolate thermal performance using a list of fits}
\usage{
interpolateAtTemp(fitList, temp)
}
\arguments{
\item{fitList}{List of model fits (output from fitModels())}

\item{temp}{Numerical value (not a vector with multiple values),
temperature value for which interpolation is to be calculated.}
}
\value{
list
}
\description{
Interpolate thermal performance using a list of fits
}

