% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fitList.R
\name{plot.fitList}
\alias{plot.fitList}
\title{Accepted parameters: xlim, ylim}
\usage{
\method{plot}{fitList}(fitList, fitCol = NULL, drawAverageFit = T, ...)
}
\arguments{
\item{fitCol}{If set to "weights", color the fit curves based on their
AIC weight}

\item{drawAverageFit}{If TRUE, draw an average fit based on the AIC weights}
}
\description{
Accepted parameters: xlim, ylim
}
